+++ Tests/StreamChatTests/APIClient/Endpoints/ReactionEndpoint_Tests.swift
@@ -7,41 +7,6 @@
 import XCTest
 
 final class ReactionEndpoints_Tests: XCTestCase {
-    func test_loadReactions_buildsCorrectly() {
-        let messageId: MessageId = "ID"
-        let pagination: Pagination = .init(pageSize: 10)
-
-        let endpoint: Endpoint<MessageReactionsPayload> = .loadReactions(
-            messageId: messageId,
-            pagination: pagination
-        )
-
-        XCTAssertEqual(endpoint.path.value, "messages/ID/reactions")
-        XCTAssertEqual(endpoint.method, .get)
-        XCTAssertTrue(endpoint.queryItems == nil)
-        XCTAssertEqual(endpoint.requiresConnectionId, false)
-        XCTAssertEqual(endpoint.body?.asAnyEncodable, pagination.asAnyEncodable)
-    }
-
-    func test_loadReactionsV2_buildsCorrectly() {
-        let messageId: MessageId = "ID"
-        let query: ReactionListQuery = .init(
-            messageId: messageId,
-            pagination: .init(pageSize: 20, offset: 0),
-            filter: .equal(.reactionType, to: "like")
-        )
-
-        let endpoint: Endpoint<MessageReactionsPayload> = .loadReactionsV2(
-            query: query
-        )
-
-        XCTAssertEqual(endpoint.path.value, "messages/ID/reactions")
-        XCTAssertEqual(endpoint.method, .post)
-        XCTAssertTrue(endpoint.queryItems == nil)
-        XCTAssertEqual(endpoint.requiresConnectionId, false)
-        XCTAssertEqual(endpoint.body?.asAnyEncodable, query.asAnyEncodable)
-    }
-
     func test_addReaction_buildsCorrectly() {
         let messageId: MessageId = .unique
         let reaction: MessageReactionType = .init(rawValue: "like")